Cleanup: int to uint8_t booleans, eliminate volatile stream pointers, fix ForceReset redundancy
DirFunction.h:
- IsDirectory/IsFinished/InSubDir: int -> uint8_t (-3 B SRAM on AVR)
  All are boolean flags assigned only 0/1; int was 2 B each needlessly.

DirFunction.cpp:
- ForceReset(): remove Prepare() call that duplicated the ResyncDirFromCwd()
  + CountEntries() work already done by ToRoot(). No functional change.

CartApi.cpp:
- Remove streamBuffer1/2 volatile static pointer variables (-4 B SRAM).
  They were permanently initialized to sharedBuf.io2.stream1/2 and never
  reassigned; replace all references with direct array access.
- DoubleBufferedStreaming ISR: replace if/else-if(usedBuffer==1) with a
  ternary. usedBuffer is strictly binary (1 - usedBuffer); the else-if
  guard was dead code.
- HandleStream: remove dead locals initialDelay and countStreamedBytes
  (received via protocol but never used in the implementation).

+++ Arduino/EasySD/DirFunction.h
@@ -56,9 +56,9 @@ class DirFunction {
     // It intentionally stores only the leading part of a filename.
     char currentFileName[32];
     uint16_t currentDirIdx; // FAT dir-entry index of the last Iterate() result
-    int IsDirectory;
-    int IsFinished;
-    int InSubDir;
+    uint8_t IsDirectory;
+    uint8_t IsFinished;
+    uint8_t InSubDir;
 };
 
 #endif  // _DIR_FUNCTION_H
